% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{residuals.modal_arima}
\alias{residuals.modal_arima}
\title{Extract Residuals from Modal ARIMA models}
\usage{
\method{residuals}{modal_arima}(object, ...)
}
\arguments{
\item{object}{An object of class \code{modal_arima}.}

\item{...}{Additional arguments.}
}
\value{
A numeric vector of residuals.
}
\description{
Extract Residuals from Modal ARIMA models
}
