% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_modal_arima.R
\name{auto.modal.arima}
\alias{auto.modal.arima}
\title{Automatic selection of Parametric Modal ARIMA model}
\usage{
auto.modal.arima(y, d = NA, max.p = 5, max.q = 5, ic = c("aic", "bic"))
}
\arguments{
\item{y}{numeric vector or time series of observations}

\item{d}{Integer, degree of differencing. If NA, it's determined automatically.}

\item{max.p}{Maximum AR order}

\item{max.q}{Maximum MA order}

\item{ic}{Information criterion to be used in model selection ("aic", "bic")}
}
\value{
An object of class \code{modal_arima}.
}
\description{
Automatic selection of Parametric Modal ARIMA model
}
\examples{
set.seed(123)
y <- arima.sim(n = 200, list(ar = 0.5))
mod <- auto.modal.arima(y, d = 0, max.p = 2, max.q = 2)
summary(mod)
}
