% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_modal_arima.R
\name{auto.modal_arima}
\alias{auto.modal_arima}
\title{Automatic selection of Parametric Modal ARIMA model}
\usage{
auto.modal_arima(y, d = NA, max.p = 5, max.q = 5, ic = c("aic", "bic"))
}
\arguments{
\item{y}{numeric vector or time series of observations}

\item{d}{Integer, degree of differencing. If NA, it's determined automatically.}

\item{max.p}{Maximum AR order}

\item{max.q}{Maximum MA order}

\item{ic}{Information criterion to be used in model selection ("aic", "bic")}
}
\value{
An object of class \code{modal_arima}.
}
\description{
Automatic selection of Parametric Modal ARIMA model
}
\details{
This function performs a grid search over AR and MA orders to find the
optimal Modal ARIMA model based on the selected information criterion (AIC or BIC).
}
\note{
GitHub repository: \url{https://github.com/chedgala/ModalForecast}
}
\examples{
\donttest{
# 1. Simulate an asymmetric AR(1) time series
set.seed(123)
y <- arima.sim(n = 200, list(ar = 0.5))

# 2. Fit a Modal ARIMA model
mod <- fit_modal_arima(y, order = c(1, 0, 0))

# 3. Model Information and AIC/BIC
summary(mod)
AIC(mod)
BIC(mod)

# 4. Run residual diagnostics with hypothesis tests
diagnostics(mod)

# 5. Produce forecasts with Asymptotic & Bootstrap prediction bands
pred_asymp <- forecast(mod, h = 10, level = c(80, 95), interval = "asymptotic")
pred_boot <- forecast(mod, h = 10, level = c(80, 95), interval = "bootstrap", npaths = 500)

# The output contains the upper and lower bounds for the respective levels:
print(pred_asymp$lower)
}
}
\references{
Fernandez, C. and Steel, M. F. J. (1998). On Bayesian Modeling of Fat Tails
and Skewness. Journal of the American Statistical Association, 93(441), 359-371.
}
\seealso{
\code{\link{fit_modal_arima}}
}
