% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modal_arima.R
\name{fit_modal_arima}
\alias{fit_modal_arima}
\title{Fit a Parametric Modal ARIMA Model using the Skew Normal Distribution}
\usage{
fit_modal_arima(y, order = c(1, 0, 0))
}
\arguments{
\item{y}{numeric vector or time series of observations}

\item{order}{A specification of the non-seasonal part of the ARIMA
model: the three components (p, d, q) are the AR order, the
degree of differencing, and the MA order.}
}
\value{
An object of class \code{modal_arima}.
}
\description{
This function fits an ARIMA(p, d, q) model where the conditional mode of the 
response is modeled using a Fernandez-Steel skew-normal distribution. 
The mode is directly parameterized by the location parameter.
}
\examples{
set.seed(123)
y <- arima.sim(n = 200, list(ar = 0.5))
mod <- fit_modal_arima(y, order = c(1, 0, 0))
summary(mod)
}
