% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{forecast.modal_arima}
\alias{forecast.modal_arima}
\title{Forecast methodology for Modal ARIMA}
\usage{
\method{forecast}{modal_arima}(object, h = 10, ...)
}
\arguments{
\item{object}{A modal_arima object.}

\item{h}{The forecast horizon.}

\item{...}{Additional arguments.}
}
\value{
A forecast object.
}
\description{
Forecast methodology for Modal ARIMA
}
