% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModalForecast-package.R
\docType{package}
\name{ModalForecast-package}
\alias{ModalForecast}
\alias{ModalForecast-package}
\title{ModalForecast: Parametric Modal ARIMA and Seasonal ARIMA Models using the SKD Family}
\description{
Implements parametric modal Autoregressive Integrated Moving Average (ARIMA) and seasonal ARIMA (SARIMA) models utilizing the Skewed Distribution (SKD) family, in which the conditional mode, rather than the conditional mean, follows the (seasonal) ARIMA recursion. Current distributions supported are the Skew-Normal, Skewed Student-t, and Skewed Laplace. The parameters are estimated by maximum likelihood using analytical gradients. Includes residual diagnostics, simulation envelopes, automatic order selection, joint and marginal modal forecasts, exact and parametric bootstrap prediction intervals, and classical asymptotic inference via the Fisher Information matrix. Methods are described in Galarza, C.E., Lachos, V.H., Cabral, C.R.B., & Castro, L.M. (2017) \doi{10.1002/sta4.140}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/chedgala/ModalForecast}
  \item Report bugs at \url{https://github.com/chedgala/ModalForecast/issues}
}

}
\author{
\strong{Maintainer}: Christian Galarza \email{chedgala@espol.edu.ec}

}
\keyword{internal}
