% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{residuals.modal_arima}
\alias{residuals.modal_arima}
\title{Residuals method for Modal ARIMA models}
\usage{
\method{residuals}{modal_arima}(object, ...)
}
\arguments{
\item{object}{an object of class \code{modal_arima}}

\item{...}{additional arguments}
}
\value{
numeric vector of residuals
}
\description{
Residuals method for Modal ARIMA models
}
